% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCredentials.R
\name{getCredentials}
\alias{getCredentials}
\title{Get  S3 credentidal from aws ec2}
\usage{
getCredentials(rolename = NULL)
}
\arguments{
\item{rolename}{string}
}
\value{
Send enviroment variabls
}
\description{
This function get S3 credentials when you are in a ec2 instance
}

